requirements.txt
langgraph==0.2.34
langchain==0.3.7
langchain-google-genai==2.0.5
python-dotenv==1.0.1
google-generativeai==0.8.3